Makes it so it has a 'content' directory but also a static directory used to host the javascript, css etc.
