+++ web/controller.go
@@ -1,14 +1,13 @@
-package main
+package web
 
 import (
     "fmt"
-    "time"
-    "flag"
     "log"
     "encoding/json"
     "net/http"
     "io/ioutil"
     "gorilla/mux"
+	"contented/utils"
 )
 
 type HttpStdResults struct{
@@ -22,35 +21,25 @@ type HttpErrResults struct{
     Debug string `json:"debug"`
 }
 
-
 var dir string
 var validDirs map[string]bool
 
-func main() {
-    flag.StringVar(&dir, "dir", ".", "Directory to serve files from")
-    flag.Parse()
+func SetupContented(router *mux.Router, contentDir string) {
+	dir = contentDir
+    validDirs = utils.GetDirectoriesLookup(dir)
 
-    fmt.Println("Using this directory As the static root: ", dir)
-
-    router := mux.NewRouter()
-    router.PathPrefix("/static/").Handler(http.StripPrefix("/static/", http.FileServer(http.Dir(dir))))
+    router.PathPrefix("/contented/").Handler(http.StripPrefix("/contented/", http.FileServer(http.Dir(dir))))
     router.HandleFunc("/content/", ListDefaultHandler)
     router.HandleFunc("/content/{dir_to_list}", ListSpecificHandler)
 
     // Host the index.html, also assume that all angular UI routes are going to be under contented
     router.HandleFunc("/", Index)
     router.HandleFunc("/contented/{path}", Index)
+}
 
-    validDirs = getDirectoriesLookup(dir)
-    srv := &http.Server{
-        Handler:      router,
-        Addr:         "127.0.0.1:8000",
-        // Good practice: enforce timeouts for servers you create!
-        WriteTimeout: 15 * time.Second,
-        ReadTimeout:  15 * time.Second,
-    }
-    log.Fatal(srv.ListenAndServe())
-    http.Handle("/", router)
+
+func SetupStatic(router *mux.Router, staticDir string) {
+    router.PathPrefix("/static/").Handler(http.StripPrefix("/static/", http.FileServer(http.Dir(staticDir))))
 }
 
 
@@ -77,7 +66,7 @@ func Index(w http.ResponseWriter, r *http.Request) {
 func ListDefaultHandler(w http.ResponseWriter, r *http.Request) {
     log.Println("Calling into ListDefault")
     w.Header().Set("Content-Type", "application/json")
-    j, _ := json.Marshal(ListDirs(dir, 3))
+    j, _ := json.Marshal(utils.ListDirs(dir, 3))
     w.Write(j)
 }
 
@@ -103,48 +92,8 @@ func getDirectory(dir string, argument string) HttpStdResults {
     path := dir + argument
     response := HttpStdResults{
         true,
-        GetDirContents(path, 1000),
+        utils.GetDirContents(path, 1000),
         "static/" + argument + "/",
     }
     return response
-}
-
-/**
- *  Check if a directory is a legal thing to view
- */
-func getDirectoriesLookup(legal string) map[string]bool {
-    var listings = make(map[string]bool)
-    files, _ := ioutil.ReadDir(legal)
-    for _, f := range files {
-        if f.IsDir() {
-            listings[f.Name()] = true
-        }
-    }
-    return listings
-}
-
-/**
- * Grab a small preview list of all items in the directory.
- */
-func ListDirs(dir string, previewCount int) map[string][]string {
-    var listings = make(map[string][]string)
-    files, _ := ioutil.ReadDir(dir)
-    for _, f := range files {
-        if f.IsDir() {
-            listings[f.Name()] = GetDirContents(dir + f.Name(), previewCount)
-        }
-    }
-    return listings
-}
-
-func GetDirContents(dir string, limit int) []string {
-    var arr = []string{}
-    imgs, _ := ioutil.ReadDir(dir)
-
-    for _, img := range imgs {
-        if !img.IsDir() && len(arr) < limit + 1 {
-            arr = append(arr, img.Name())
-        }
-    }
-    return arr
 }
